Created separate component for tabs
